maint(all) Upgrade to scala 2.13 and 2.6.21

+++ prometheus/src/main/scala/filodb/prometheus/parse/AntlrParser.scala
@@ -468,7 +468,7 @@ class AntlrParser extends PromQLBaseVisitor[Object] {
     } else {
       val seconds = getTotalSecondsFromAntlrDurationString(str)
       assert(seconds > 0, "bracket-notation duration cannot be zero")
-      Duration(seconds, Second)
+      Duration(seconds.toDouble, Second)
     }
   }
 }
